feat: upgrade sqlness to latest version
