+++ tests/Makefile
@@ -7,7 +7,7 @@ export CERESDB_BINARY_PATH ?= $(ROOT)/../target/$(MODE)/ceresdb-server
 export CERESDB_CONFIG_PATH ?= $(ROOT)/../docs/sqlness.toml
 export CERESDB_SERVER_ENDPOINT ?= 127.0.0.1:8831
 export CERESDB_TEST_CASE_PATH ?= $(ROOT)/cases
-export CERESDB_TEST_SQLNESS ?= $(ROOT)/../target/$(MODE)/ceresdb-test
+export CERESDB_TEST_SQLNESS ?= $(ROOT)/../target/$(MODE)/ceresdb-integration-test
 export CERESDB_STDOUT_FILE ?= /tmp/ceresdb-stdout.log
 export CERESDB_STDERR_FILE ?= /tmp/ceresdb-stderr.log
