+++ tests/sqlness/Cargo.toml
@@ -9,5 +9,5 @@ workspace = true
 anyhow = "1.0.58"
 async-trait = "0.1"
 ceresdb-client-rs = { git = "https://github.com/CeresDB/ceresdb-client-rs.git", rev = "a9d9190f4f7b55171ea2ad142fb41dc9909c19c5" }
-sqlness = { git = "https://github.com/CeresDB/sqlness.git", rev = "c077b17d73ab25460c152dc34e8f80f904522a57" }
+sqlness = { git = "https://github.com/CeresDB/sqlness.git", rev = "0f0143a8df319e19be8411e6bd54330985a684d6" }
 tokio = { workspace = true }
