+++ tests/sqlness/src/main.rs
@@ -1,29 +1,39 @@
 // Copyright 2022 CeresDB Project Authors. Licensed under Apache-2.0.
 
-#![feature(try_blocks)]
-
 use std::env;
 
 use anyhow::Result;
-use sqlness::Runner;
+use async_trait::async_trait;
+use database::CeresDB;
+use sqlness::{EnvController, Runner};
 
-mod client;
-mod setup;
+mod database;
 
 const CASE_ROOT_PATH_ENV: &str = "CERESDB_TEST_CASE_PATH";
+
+pub struct CeresDBController;
+
+#[async_trait]
+impl EnvController for CeresDBController {
+    type DB = CeresDB;
+
+    async fn start(&self, _mode: &str, config: Option<String>) -> Self::DB {
+        CeresDB::try_new(config)
+    }
+
+    /// Stop one [`Database`].
+    async fn stop(&self, _mode: &str, database: Self::DB) {
+        database.stop();
+    }
+}
 
 #[tokio::main]
 async fn main() -> Result<()> {
     let case_dir = env::var(CASE_ROOT_PATH_ENV)?;
-    let env = setup::CeresDBEnv::start_server();
-    let config = sqlness::Config {
-        case_dir,
-        test_case_extension: String::from("sql"),
-        output_result_extension: String::from("output"),
-        expect_result_extension: String::from("result"),
-        interceptor_prefix: String::from("-- SQLNESS"),
-        env_config_file: String::from("config.toml"),
-    };
+    let env = CeresDBController;
+    let config = sqlness::ConfigBuilder::default()
+        .case_dir(case_dir)
+        .build()?;
     let runner = Runner::new_with_config(config, env).await?;
     runner.run().await?;
